Reduce the maximum length of slug fields to 752 characters
Recent versions of mysql (tested with 8.0.19) limit the size of key length to
3072 bytes, effectively limiting the size of varchar indices to 767 characters
when using utf8mb4 encoding. Commit 747ec236868820fe662176777e15a930680ebfe0
sets the minimum length of slug fields to 767 characters to address the
issue. However, when slug fields appear in composite indices, the limit of 767
characters applies to the _composite_ index and not just the slug part. This
limits the maximum length even further.

With freshly installed database running migrations leads to this:

$ python manage.py migrate
  Apply all migrations: ...
Running migrations:
  ...
  Applying djangocms_blog.0032_auto_20180109_0023...Traceback (most recent call last):
  ...
  django.db.utils.OperationalError: (1071, 'Specified key was too long; max key length is 3072 bytes')

This is due to slug field in `Post` and `BlogCategory` models appearing in
composite index with `language_code`. `language_code` has type varchar(15),
effectively limiting the maximum lenght of `slug` to 752 characters (752+15=767)
to ensure successful migrations.

This commit reduces maximum length of slug fields to 752 characters. It was done
to existing migrations. While this may cause inconsistencies for old
installations that have already applied those migrations, it's necessary for
fresh installations using utf8mb4 to apply the migrations in the first place.